[Tests] Move some types to IdeUnitTests for reusability
