+++ main/tests/Ide.Tests/MonoDevelop.Ide.Tests.csproj
@@ -124,13 +124,8 @@
     <Compile Include="MonoDevelop.SourceEditor\DebugTooltipTests.cs" />
     <Compile Include="MonoDevelop.SourceEditor\JSonIndentEngineTests.cs" />
     <Compile Include="MonoDevelop.SourceEditor\SourceEditorTests.cs" />
-    <Compile Include="IdeTestBase.cs" />
     <Compile Include="MonoDevelop.Ide\UserPreferencesTests.cs" />
     <Compile Include="MonoDevelop.Ide.Projects\ResourcesTests.cs" />
-    <Compile Include="MonoDevelop.Ide.Gui\TestViewContent.cs" />
-    <Compile Include="MonoDevelop.Ide.Gui\TestWorkbenchWindow.cs" />
-    <Compile Include="MonoDevelop.Ide.Gui\TestDocument.cs" />
-    <Compile Include="TypeSystemServiceTestExtensions.cs" />
     <Compile Include="MonoDevelop.Ide.Editor\SkipCharSessionTests.cs" />
     <Compile Include="MonoDevelop.Components\PangoUtilTests.cs" />
     <Compile Include="MonoDevelop.Ide.Projects\DotNetCoreProjectTests.cs" />
@@ -141,7 +136,6 @@
     <Compile Include="MonoDevelop.Ide.Tasks\CommentTasksProviderTests.cs" />
     <Compile Include="MonoDevelop.Ide.Tasks\CommentTasksProviderTests.Controller.cs" />
     <Compile Include="MonoDevelop.Ide.Composition\CompositionManager.CachingTests.cs" />
-    <Compile Include="TextEditorExtensionTestBase.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices\MonoDevelopProjectCacheHostServiceTests.cs" />
     <Compile Include="MonoDevelop.Ide\RoslynSearchCategoryTests.cs" />
     <Compile Include="MonoDevelop.Ide.Editor\CodeSnippetTests.cs" />
@@ -153,6 +147,7 @@
     <Compile Include="MonoDevelop.Ide.RoslynServices.Options\MonoDevelopOptionPersisterIntegrationTests.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices.Options\MockOptionService.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices.Options\MonoDevelopOptionPersisterTests.cs" />
+    <Compile Include="TypeForwarders.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
@@ -208,6 +203,10 @@
       <Project>{F8F92AA4-A376-4679-A9D4-60E7B7FBF477}</Project>
       <Name>MonoDevelop.SourceEditor</Name>
       <Private>False</Private>
+    </ProjectReference>
+    <ProjectReference Include="..\IdeUnitTests\IdeUnitTests.csproj">
+      <Project>{F7B2B155-7CF4-42C4-B5AF-63C0667D2E4F}</Project>
+      <Name>IdeUnitTests</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
