+++ main/tests/IdeUnitTests/IdeTestBase.cs
@@ -1,10 +1,10 @@
 ﻿//
-// MyClass.cs
+// IdeTestBase.cs
 //
 // Author:
-//       Marius Ungureanu <maungu@microsoft.com>
+//       Lluis Sanchez <llsan@microsoft.com>
 //
-// Copyright (c) 2018 Microsoft Inc.
+// Copyright (c) 2017 Microsoft
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -23,13 +23,19 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-using System;
-namespace IdeUnitTests
+using UnitTests;
+
+namespace MonoDevelop.Ide
 {
-	public class MyClass
+	public class IdeTestBase: RoslynTestBase
 	{
-		public MyClass ()
+		protected override void InternalSetup(string rootDir)
 		{
+			base.InternalSetup(rootDir);
+
+			Xwt.Application.Initialize(Xwt.ToolkitType.Gtk);
+			Gtk.Application.Init();
+			DesktopService.Initialize();
 		}
 	}
 }
