+++ main/tests/IdeUnitTests/Makefile.am
@@ -1,0 +1,1 @@
+include $(top_srcdir)/xbuild.include
