+++ main/tests/IdeUnitTests/TestViewContent.cs
