+++ main/tests/IdeUnitTests/TestWorkbenchWindow.cs
