+++ main/tests/IdeUnitTests/TextEditorExtensionTestBase.cs
