+++ main/tests/MacPlatform.Tests/MacPlatform.Tests.csproj
@@ -75,15 +75,17 @@
     <ProjectReference Include="..\UnitTests\UnitTests.csproj">
       <Project>{1497D0A8-AFF1-4938-BC22-BE79B358BA5B}</Project>
       <Name>UnitTests</Name>
+      <Private>False</Private>
     </ProjectReference>
     <ProjectReference Include="..\MonoDevelop.Core.Tests\MonoDevelop.Core.Tests.csproj">
       <Project>{FDA43CAA-1C2A-4593-8601-3E2EE06D9E03}</Project>
       <Name>MonoDevelop.Core.Tests</Name>
       <Private>False</Private>
     </ProjectReference>
-    <ProjectReference Include="..\Ide.Tests\MonoDevelop.Ide.Tests.csproj">
-      <Project>{73D4CC8B-BAB9-4A29-841B-F25C6311F067}</Project>
-      <Name>MonoDevelop.Ide.Tests</Name>
+    <ProjectReference Include="..\IdeUnitTests\IdeUnitTests.csproj">
+      <Project>{F7B2B155-7CF4-42C4-B5AF-63C0667D2E4F}</Project>
+      <Name>IdeUnitTests</Name>
+      <Private>False</Private>
     </ProjectReference>
   </ItemGroup>
 </Project>
